Pulls out service maturity report
