+++ src/components/EntityOpsLevelMaturityContent.tsx
@@ -6,17 +6,14 @@ import Alert from "@material-ui/lab/Alert";
 import { useApi } from "@backstage/core-plugin-api";
 import { useEntity } from "@backstage/plugin-catalog-react";
 import { useAsync, useAsyncFn } from "react-use";
-import { cloneDeep } from "lodash";
-import EntityOpsLevelMaturityProgress from "./EntityOpsLevelMaturityProgress";
 import { opslevelApiRef } from "../api";
+import ServiceMaturityReport from "./ServiceMaturityReport";
 import {
   LevelCategory,
   OpsLevelServiceData,
   ScorecardStats,
 } from "../types/OpsLevelData";
 import { SnackAlert, SnackbarProps } from "./SnackAlert";
-import CheckResultsByLevel from "./CheckResultsByLevel";
-import ServiceMaturitySidebar from "./ServiceMaturitySidebar";
 
 // eslint-disable-next-line import/prefer-default-export -- One of the top level exports
 export function EntityOpsLevelMaturityContent() {
@@ -31,9 +28,7 @@ export function EntityOpsLevelMaturityContent() {
     message: "",
     severity: "info",
   });
-  const [selectedCategories, setSelectedCategories] = useState<Array<String>>(
-    [],
-  );
+  const [selectedCategories, setSelectedCategories] = useState<string[]>([]);
   const [fetchState, doFetch] = useAsyncFn(async () => {
     const result = await opslevelApi.getServiceMaturityByAlias(serviceAlias);
 
@@ -74,14 +69,6 @@ export function EntityOpsLevelMaturityContent() {
   }
 
   const service = opsLevelData?.account?.service;
-  let opslevelUrl = "https://app.opslevel.com";
-  if (service) {
-    const extensionToRemove = `/services/${entity.metadata.name}`;
-    opslevelUrl = service.htmlUrl
-      .split(extensionToRemove)
-      .filter((s) => s)
-      .join(extensionToRemove);
-  }
 
   function showSnackbar(snackbarProps: SnackbarProps) {
     setSnackbarOpen(true);
@@ -171,109 +158,12 @@ export function EntityOpsLevelMaturityContent() {
   }
 
   const { maturityReport } = service;
-  const levels = opsLevelData.account.rubric?.levels?.nodes;
-  const levelCategories =
-    opsLevelData.account.service.maturityReport?.categoryBreakdown.map((c) => {
-      return { ...c, rollsUp: true };
-    }) || [];
+  const levels = opsLevelData.account.rubric.levels.nodes;
   const checkResultsByLevel =
     opsLevelData.account.service.serviceStats?.rubric?.checkResults?.byLevel
       ?.nodes;
   const scorecards =
     opsLevelData.account.service.serviceStats?.scorecards?.nodes;
-  const scorecardCategories = scorecards
-    ?.reduce((scoreCardAccumulator: LevelCategory[], currentScorecard) => {
-      if (!currentScorecard.categories?.edges) {
-        return scoreCardAccumulator;
-      }
-      return [
-        ...scoreCardAccumulator,
-        ...currentScorecard.categories.edges.reduce(
-          (nodeAccumulator: LevelCategory[], currentEdge): LevelCategory[] => [
-            ...nodeAccumulator,
-            {
-              category: {
-                id: currentEdge.node?.id ?? "",
-                name: currentEdge.node?.name ?? "",
-              },
-              level: currentEdge.level?.name
-                ? { name: currentEdge.level?.name }
-                : null,
-              rollsUp: currentScorecard?.scorecard?.affectsOverallServiceLevels,
-            },
-          ],
-          [],
-        ),
-      ];
-    }, [])
-    .sort((a, b) => (a.category.name < b.category.name ? -1 : 1));
-
-  function checksByLevelIncludingScorecards() {
-    if (!checkResultsByLevel) {
-      return [];
-    }
-    const result = cloneDeep(checkResultsByLevel);
-
-    result.forEach((checkResults) => {
-      // Use level's 'index' field b/c we don't have level ID here. Note: 'index' *is not* the same as array index
-      const levelIndex = checkResults.level.index;
-      scorecards?.forEach((scorecard) => {
-        const entry = scorecard.checkResults?.byLevel?.nodes?.find(
-          (node) => node.level.index === levelIndex,
-        );
-        if (!entry) {
-          return;
-        }
-
-        entry.items.nodes.forEach((node) => {
-          // eslint-disable-next-line no-param-reassign -- This is taken from OpsLevel and keeping them in sync should be prioritized
-          node.check.isScorecardCheck = true;
-        });
-        // eslint-disable-next-line no-param-reassign -- This is taken from OpsLevel and keeping them in sync should be prioritized
-        checkResults.items.nodes = [
-          ...checkResults.items.nodes,
-          ...entry.items.nodes,
-        ].filter(
-          (i) =>
-            !!i.check.category &&
-            selectedCategories.includes(i.check.category.id),
-        );
-      });
-    });
-    return result;
-  }
-
-  const allCheckResultsByLevel = checksByLevelIncludingScorecards();
-
-  function totalChecks() {
-    return allCheckResultsByLevel?.reduce(
-      (accumulator, checkByLevel) =>
-        accumulator + checkByLevel.items.nodes.length,
-      0,
-    );
-  }
-
-  const ResultStatusEnum = Object.freeze({
-    FAILED: "failed",
-    PENDING: "pending",
-    PASSED: "passed",
-    UPCOMING_FAILED: "upcoming_failed",
-    UPCOMING_PENDING: "upcoming_pending",
-    UPCOMING_PASSED: "upcoming_passed",
-  });
-
-  function totalPassingChecks() {
-    return allCheckResultsByLevel.reduce(
-      (accumulator, checkByLevel) =>
-        accumulator +
-        checkByLevel.items.nodes.reduce(
-          (passingChecks, check) =>
-            passingChecks + (check.status === ResultStatusEnum.PASSED ? 1 : 0),
-          0,
-        ),
-      0,
-    );
-  }
 
   if (!maturityReport) {
     return (
@@ -285,103 +175,21 @@ export function EntityOpsLevelMaturityContent() {
       />
     );
   }
-  function updateCategoryIdSelection(
-    addedCategoryIds: Array<String>,
-    removedCategoryIds: Array<String>,
-  ) {
-    const newSelection = selectedCategories.filter(
-      (c) => !removedCategoryIds.includes(c),
-    );
-    addedCategoryIds.forEach((id) => {
-      if (!newSelection.includes(id)) {
-        newSelection.push(id);
-      }
-    });
-    setSelectedCategories(newSelection);
-  }
-
-  const overallLevel = (() => {
-    const sortedLevels = levels.sort((a, b) => (a.index > b.index ? 1 : -1));
-    const sortedCheckResults = allCheckResultsByLevel.sort((a, b) =>
-      a.level.index > b.level.index ? 1 : -1,
-    );
-
-    let i;
-    for (i = 0; i < sortedCheckResults.length; i++) {
-      if (
-        sortedCheckResults[i].items.nodes.some(
-          (e) => e.status === "failed" || e.status === "pending",
-        )
-      ) {
-        return sortedLevels[i];
-      }
-    }
-    return sortedLevels[i];
-  })();
-
-  function ServiceMaturityReport() {
-    return (
-      <>
-        <Grid container item xs={12} justifyContent="flex-end" direction="row">
-          <Grid item>
-            <Button
-              variant="contained"
-              color="primary"
-              onClick={exportEntity}
-              disabled={exporting}
-            >
-              Update Entity
-            </Button>
-          </Grid>
-          <Grid item>
-            <Button
-              variant="contained"
-              color="primary"
-              target="_blank"
-              href={`${service?.htmlUrl}/maturity-report`}
-            >
-              View Maturity in OpsLevel
-            </Button>
-          </Grid>
-        </Grid>
-        <Grid container>
-          <Grid item xs={4}>
-            {maturityReport?.overallLevel && (
-              <ServiceMaturitySidebar
-                levels={levels}
-                scorecardCategories={scorecardCategories}
-                levelCategories={levelCategories}
-                overallLevel={maturityReport.overallLevel}
-                selectedCategoryIds={selectedCategories}
-                onCategorySelectionChanged={updateCategoryIdSelection}
-              />
-            )}
-          </Grid>
-          <Grid container item xs={8} direction="column">
-            <Grid item>
-              <EntityOpsLevelMaturityProgress
-                levels={levels}
-                serviceLevel={overallLevel}
-              />
-            </Grid>
-            <Grid item>
-              <CheckResultsByLevel
-                opslevelUrl={opslevelUrl}
-                checkResultsByLevel={allCheckResultsByLevel}
-                totalChecks={totalChecks()}
-                totalPassingChecks={totalPassingChecks()}
-              />
-            </Grid>
-          </Grid>
-        </Grid>
-      </>
-    );
-  }
 
   return (
     <Grid container spacing={2} direction="column">
       <SnackAlert {...snackbar} open={snackbarOpen} setOpen={setSnackbarOpen} />
-      <ServiceMaturityReport />
+      <ServiceMaturityReport
+        exportEntity={exportEntity}
+        exporting={exporting}
+        levels={levels}
+        service={service}
+        checkResultsByLevel={checkResultsByLevel}
+        selectedCategories={selectedCategories}
+        scorecards={scorecards}
+        overallLevel={maturityReport?.overallLevel}
+        setSelectedCategories={setSelectedCategories}
+      />
     </Grid>
   );
 }
